Implement CartContext for centralized cart management; refactor components to utilize context for cart operations and item count
