+++ src/components/cards/ItemCard.jsx
@@ -1,6 +1,8 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import CartContext from '../../context/CartContext';
 
 const ItemCard = ({ product, onViewDetails}) => {
+  const { addToCart } = useContext(CartContext)
   return (
     <div className="group relative bg-white rounded-xl shadow-md hover:shadow-2xl transition-all duration-300 overflow-hidden border border-gray-100 flex flex-col h-full">
       <div className="relative h-64 w-full overflow-hidden bg-gray-50">
@@ -53,6 +55,7 @@ const ItemCard = ({ product, onViewDetails}) => {
             ${product.availability 
                 ? 'bg-gray-100 hover:bg-blue-600 hover:text-white text-gray-800' 
                 : 'bg-gray-100 text-gray-300 cursor-not-allowed'}`}
+            onClick={() => {addToCart(product._id, 1)}}
           >
             <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" viewBox="0 0 16 16">
               <path d="M0 1.5A.5.5 0 0 1 .5 1H2a.5.5 0 0 1 .485.379L2.89 3H14.5a.5.5 0 0 1 .491.592l-1.5 8A.5.5 0 0 1 13 12H4a.5.5 0 0 1-.491-.408L2.01 3.607 1.61 2H.5a.5.5 0 0 1-.5-.5zM5 12a2 2 0 1 0 0 4 2 2 0 0 0 0-4zm7 0a2 2 0 1 0 0 4 2 2 0 0 0 0-4zm-7 1a1 1 0 1 1 0 2 1 1 0 0 1 0-2zm7 0a1 1 0 1 1 0 2 1 1 0 0 1 0-2z"/>
